Caught requests for iOS 7's wispr service
iOS 7 changes the wispr service used to validate that a device has a
valid network connection. It now hits the host www.appleiphonecell.com
and expects the same response as the old service returned. This
condition has now been added so that the newer OS will believe it had a
valid connection after being Pineappled.